requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Cython==0.29.37
exceptiongroup==1.2.0
frozenlist==1.4.1
future==0.18.3
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
naverpaper==0.1
packaging==23.2
playwright==1.40.0
pluggy==1.3.0
pyasn1==0.5.1
pyc==0.1.0
pyee==11.0.1
pytest==7.4.4
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-dotenv==1.0.0
python-slugify==8.0.1
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
